refactor: integrate ChunkedProcessor into speckle filters
- OTB/CMake improvements: GDAL wrapper fixes, Boost config updates
- Classification: KMeans, NormalBayes, SVM classifier optimizations
- Georeferencing: GCP transformer, RPC transformer, image warper improvements
- Segmentation: segment features, segment map, simple segmenter enhancements
- UI: main window, log panel, histogram widget, spectral profile improvements
- Processing: chunked processor, mosaic, change detection, image enhancement
- Tests: comprehensive test suite additions
- All 435 tests pass with QT_QPA_PLATFORM=offscreen

+++ src/analysis/segmentation/rs_simple_segmenter.h
@@ -38,7 +38,7 @@ class QGIS_ANALYSIS_EXPORT RsSimpleSegmenter
     static void gaussianSmooth( QVector<float> &data, int w, int h, int kernelSize );
 
     /// Quantize float data to integer bins.
-    static QVector<int> quantize( const float *data, int n, int bins, float nodata );
+    static QVector<int> quantize( const float *data, size_t n, int bins, float nodata );
 
     /// 8-connected component labeling on quantized integer grid.
     /// Returns label map (1-based; 0 = background/nodata).
